Add semantic versioning on every git tag push to master (#1)
* Create a new semver release on git tag push

Go modules dictates[1] that we should publish a new semantic version[2]
for every trunk commit (even if our module is not intended for use by
others).  This is a good thing.

Use GoReleaser[3] to create a new release every time a Git tag is pushed
to GitHub.

[1] https://blog.golang.org/publishing-go-modules
[2] https://semver.org/
[3] https://goreleaser.com/ci/

* Skip building of executable when releasing

As we don't have anything to build yet.