Fix bug where consumers were unable to merge PRs (#14)
Fixes a bug whereby a PR could show that the check has passed, but
still not allow the pull request to be merged.  This is because GitHub
Actions create a separate new instance of the check for some different
event types, e.g. the `labeled` event creates a new check instance
separate from the `push` event checks.  So if, for instance, a consumer:
1. pushed a commit when there is no required label
2. added the required label
then the `lableled` instance of the check would pass, but the push would
still show the initial failure.
The fix in this commit is to retrieve the GitHub
labels for the pull request via the GitHub API, instead of using the
labels in the GitHub Action's event JSON (which is a frozen snapshot of
the labels at the time of committing and therefore does not stay up to
date with subsequent label changes).  Now if we rerun the failed `push`
job after amending the labels, the check passes.