Tidy the input variables
Removed the unnecessary manual translation of the input variables into
environment variables, in the action.yml.  It is not needed because
GitHub actions automatically create environment variables for the input
variables.[1]

NB The manual translation would not work when we move to a minimal
pre-built Docker image (#22).

Also renamed the inputs to have underscores instead of dashes, as
dashes are not valid POSIX standard names.[2]
(This caused a problem as I could not set these environment variables
when running the tests, causing the tests to fail.)
(I fed back to GitHub (via their feedback form [3]) that they may want
to convert dashes to underscores when converting the input variables to
environment variables.)

[1] https://help.github.com/en/actions/creating-actions/metadata-syntax-for-github-actions#inputs
[2] https://stackoverflow.com/a/36992531
[3] https://support.github.com/contact/feedback?contact%5Bcategory%5D=actions