Speed up label checking by stripping symbols
This reduces the binary size from 2.7MB to 2.57MB (not a huge drop, but
worth doing nonetheless)

Using the strip command as per this example:
https://www.sethvargo.com/writing-github-actions-in-go/

This post[1] suggests that the Go team do not recommend doing this, as
it may break the executable.  Still going ahead though, we can always
revert this change if we encounter any problems.

[1] https://stackoverflow.com/questions/30005878/avoid-debugging-information-on-golang#comment48168448_30007760