Move settings which are not container-specific
The guideline for`devcontainer.json` settings is that they should only
contain settings which _must_ be be changed in a container (e.g.
absolute paths)[1].

Moved the other settings to `.vscode/settings.json` so that they are
more visible, and easier to use when working locally (i.e. not in a
container).

[1] https://github.com/microsoft/vscode-remote-release/issues/874#issuecomment-508135534