Use JSON simulated GitHub responses, not gzip
The original captured GitHub API responses were gzip encoded, which made
them unreadable (without manually decoding them).

There is a feature request to have Hoverfly to do this translation
automatically when exporting captured simulations, but it is not
available yet[1].

So this commit has manually placed the decoded JSON in the simulated
responses.

[1] https://github.com/SpectoLabs/hoverfly/issues/787