Fix bug which caused all checks to always run (#230)
This is an important bug fix, as prior to this commit all four label
checks (`one_of`, `none_of`, `all_of`, `any_of`) were always running,
even when not all of the checks were specified to run in the user's
GitHub Actions yaml file.  This in turn was causing unexpected results
when running the Action, e.g. it could cause the pull request validation
to fail when it should pass, and vice versa.

The bug was caused by the code expecting the check values by default
not to be set at all, whereas in reality they default to an empty
string.

As well as containing the bug fix, this commit amends the integration
tests to cover this scenario.