Enable required status checks on forked PRs (#265)
Prior to this commit the [status checks][1] that we need to run on pull
requests were only set to run on pushes.  This meant that they were not
being run when a contributor submitted a PR from a fork.  To fix this
we now run the required status check workflows explicitly on pull
requests as well as pushes.

[1]: https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/collaborating-on-repositories-with-code-quality-features/about-status-checks