Fix build on forks
The build on forks was failing because the GitHub Action which tests the
building and pushing the of the Docker image requires access to GitHub
Actions secrets, and forks rightly don't have access to them.  So this
commit [excludes forks from running this GitHub Action][1] (they don't
need to run it, as the additional confidence from it is just a nice to
have).

This ideally would have been part of #274, but was missed.

[1]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-run-job-for-specific-repository