Skip integration tests on forks
The build on forks was failing because the GitHub Action which runs the
integration tests requires a GitHub token with access to an agilepathway
repo, and forks rightly don't have access.  So this commit [excludes
forks from running this GitHub Action][1] (they don't need to run it,
as the additional confidence from it is just a nice to have, as we also
have the tests which run using mocks and which do run fine on forks).

This ideally would have been part of #294 or indeed #274, but was
missed.

[1]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-run-job-for-specific-repository