Add GitHub Enterprise tests to CI/CD pipelines (#300)
The only reason that the tests were not added as part of the previous
commit (which enabled GitHub Enterprise) was that the previous commit
was a contribution from a fork, and it was not possible for me to amend
the GitHub Actions workflow files on a fork (rightly, for security
reasons).

NB We do not run the integration tests in GitHub Enterprise Server mode
as that would require having a real GitHub Enterprise Server, which
would be expensive in time and money to maintain.  But having them run
as integration tests in GitHub Enterprise Cloud mode gives us pretty
good test coverage.