Specify GitHub Enterprise input var in action.yml
When we added support for GitHub Enterprise in the two commits
immediately preceding this one, we forgot to specify the new [GitHub
Actions input parameter][1] in the `action.yml`.  My guess is that
omission should have prevented the new GitHub Enterprise functionality
from working, but strangely when I end to end tested it manually with
a real pull request on a test repo it seemed to work ok.  Either way
this commit adds the input var to `action.yml` as it should be.

[1]: https://docs.github.com/en/actions/creating-actions/metadata-syntax-for-github-actions#inputs