Add allow failure mode
Allow the label checker to not fail the build, even when a label
check succeeds, and then use an [`if` condition][1] in a subsequent step
or job to conditionally do something or not. For example: check if a
pull request has a `preview` label and then only deploy an ephemeral
environment if so.

The `allow_failure` mode is set via an input parameter in the
[`with` section][2]: `allow_failure: true`

[1]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstepsif
[2]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstepswith