Update deprecated set-output GitHub Action command
As per the [GitHub blog article][1] about this.

Also deleted the integration tests relating to the [allow failure][2]
mode, as the recently added end-to-end tests offer much better test
coverage (the integration tests failed to catch our recent bug there,
whereas the end-to-end tests would have caught it).  So there was no
benefit in refactoring those integration tests as part of this commit,
better to delete them.

[1]: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/