Remove dependency on Mage (#332)
[Mage][1] is a Go build tool which we were using to invoke our label
checking Go function and to build our Go binary. We don't need Mage
though, as we can simply use the standard Go `main` function and build
the binary using the standard `go build` command.

It's always good to remove unneeded dependencies, and also Mage was
responsible for setting the beginning of the error output (`Error:`)
that is displayed on the console.  By removing Mage we now have control
of this.  This paves the way for making the error output a [GitHub
Actions  workflow command][2] in a future PR, which will mean the error
message is highlighted (which will fix #330).

[1]: https://magefile.org/
[2]: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-error-message