Add prefix mode to label checks
This is a new feature that checks for the presence or absence of labels
starting with a given prefix.

This prefix label checking is a powerful feature that allows pull
requests to have scoped labels [in a similar way to GitLab][1].

To use, set the `prefix_mode` as an input parameter in the
`with` section: `prefix_mode: true`.
It is an optional parameter as it defaults to false.

Any string can be used as the prefix.

`one_of`, `any_of`, `none_of` can be used when checking prefixes, but
`all_of` cannot (as `all_of` does not make sense when checking for
labels starting with a given prefix).

Only one prefix can be specified at a time. For instance
`one_of: "type:","visibility/"` is not allowed.

Closes #328

[1]: https://docs.gitlab.com/ee/user/project/labels.html#scoped-labels