Do not pin devcontainer linux dependencies anymore (#348)
The benefit of having pinned devcontainer linux dependencies is that it
ensures that the devcontainers are idempotent. But in practice not
having this is not likely to become a problem. If it ever does, we can
pin any troublesome dependencies again (it will be unlikely that we will
have to, though).