Pin major versions of GitHub Actions dependencies (#359)
Prior to this commit we were pinning to patch level dependencies. Just
pinning to major versions will make the application easier to maintain -
fewer updates.  We have sufficiently high test coverage to feel we can
make this change safely.  If it causes a problem in the future (which is
unlikely) we can pin specific dependencies at minor or patch level again
if needed.

We have left some of the dependencies still pinned to patch versions,
for instance the tag and release dependencies, as we need to be very
careful with them. We will investigate those dependencies more closely
at some time in the future and may pin them just to major versions in a
new pull request if we think it's safe enough.