Check Docker image is correctly tagged latest
A recent change, #405, which added ARM64 Docker images to the build
process, introduced a bug (see #419) whereby the new Docker images for
each release are no longer being tagged `latest`, as they should be.
This commit adds a test to catch this bug, and we will then fix the bug
in a subsequent commit.