Fix broken ARM deployments
[Arm][1] deployments have been broken since the previous release to this
one, [v.1.62][1]. The breakage wasn't caused by anything in the previous
release itself, but rather by the virtualized [QEMU][3] Docker build and
push process having developed a fault.  This was always something in
danger of happening, as the QEMU virtualized approch is not as stable
as a non virtualized approach.  Fortunately GitHub Actions [Arm runners
are now  available for free in GitHub Actions as of January 16,
2025][4]. So this fix simply replaces QEMU with native ARM runners for
the Docker build and push on Arm.

Fixes #480

[1]: https://en.wikipedia.org/wiki/ARM_architecture_family
[2]: https://github.com/agilepathway/label-checker/releases/tag/v1.6.62
[3]: https://github.com/docker/setup-qemu-action
[4]: https://github.blog/changelog/2025-01-16-linux-arm64-hosted-runners-now-available-for-free-in-public-repositories-public-preview/